update starting stats of characters

+++ app/core/include/common.h
@@ -14,8 +14,10 @@ struct Stats{
     int m_powMag = 0;
     int m_aggro = 0;
     int m_speed = 0;
+    // critical strike in %
     int m_CriticalStrike  = 0;
-    int m_dogde = 0;
+    // dodge in %
+    float m_dogde = 0;
     int m_regenHP = 0;
     int m_regenMana = 0;
     int m_regenVigueur = 0;
